+++ CryptoTrader/Trading/Traders/Simple/Strategies/BTS/Models/TraderBTSSavedState.swift
@@ -3,8 +3,8 @@ import Foundation
 struct TraderBTSSavedState: Codable {
     var openSellOperations: [TraderBTSSellOperation]
     var closedSellOperations: [TraderBTSSellOperation]
-    var currentBalance: Decimal
-    var initialBalance: Decimal
-    var profits: Decimal
+    var currentBalance: Double
+    var initialBalance: Double
+    var profits: Double
     var startDate: Date
 }
